requirements.txt
mido>=1.3.3
python-rtmidi>=1.5.8
python-dotenv
paho-mqtt
pydantic-settings
rpi-lgpio